Add NN support for monochromatic sensors

+++ robopipe_api/camera/pipeline/nn_pipeline.py
@@ -41,41 +41,61 @@ class NNPipeline(StreamingPipeline):
                 self.neural_networks[camera.getBoardSocket().name] = neural_network
                 break
 
-    def add_nn(self, nn: CameraNNConfig):
-        sensor_name = nn.sensor.name
+    def __setup_mono_camera(
+        self,
+        camera: dai.node.MonoCamera,
+        nn: CameraNNConfig,
+        nn_node: dai.node.NeuralNetwork,
+    ):
+        sensor_name = camera.getBoardSocket().name
+        script = self.scripts[sensor_name]
 
-        if sensor_name in self.neural_networks:
-            nn_node = self.neural_networks[sensor_name]
-            nn.configure_node(nn_node)
-            return
+        script.outputs["preview"].link(nn_node.input)
 
-        if sensor_name not in self.cameras:
-            cam = self.pipeline.createColorCamera()
-            cam.setBoardSocket(nn.sensor)
-            self.cameras[sensor_name] = cam
-        else:
-            cam = self.cameras[sensor_name]
+    def __setup_camera(
+        self,
+        camera: dai.node.ColorCamera | dai.node.Camera,
+        nn: CameraNNConfig,
+        nn_node: dai.node.NeuralNetwork,
+    ):
+        sensor_name = camera.getBoardSocket().name
+        camera.setPreviewSize(nn.input_shape[:2])
+        camera.setInterleaved(False)
 
-        cam.setPreviewSize(nn.input_shape[:2])
-        cam.setInterleaved(False)
-
-        nn_node = nn.create_node(self.pipeline)
-        self.neural_networks[sensor_name] = nn_node
-
-        cam_nn_out = self.create_x_link(
-            sensor_name, PipelineQueueType.NN, False, False, 1
-        )
         still_queue = self.outputs.get(
             PipelineQueueType.STILL.get_queue_name(sensor_name)
         )
 
         if still_queue is not None:
-            cam.preview.unlink(still_queue.input)
-            cam.still.link(still_queue.input)
-            cam.setNumFramesPool(2, 2, 3, 3, 2)
+            camera.preview.unlink(still_queue.input)
+            camera.still.link(still_queue.input)
+            camera.setNumFramesPool(2, 2, 3, 3, 2)
 
-        cam.preview.link(nn_node.input)
-        nn_node.out.link(cam_nn_out.input)
+        camera.preview.link(nn_node.input)
+
+    def add_nn(self, nn: CameraNNConfig):
+        sensor_name = nn.sensor.socket.name
+
+        if sensor_name in self.neural_networks:
+            nn_node = self.neural_networks[sensor_name]
+            nn.configure_node(nn_node)
+        else:
+            nn_node = nn.create_node(self.pipeline)
+            self.neural_networks[sensor_name] = nn_node
+            cam_nn_out = self.create_x_link(
+                sensor_name, PipelineQueueType.NN, False, False, 1
+            )
+            nn_node.out.link(cam_nn_out.input)
+
+        if sensor_name not in self.cameras:
+            self.add_sensor(nn.sensor)
+
+        cam = self.cameras[sensor_name]
+
+        if isinstance(cam, dai.node.MonoCamera):
+            self.__setup_mono_camera(cam, nn, nn_node)
+        else:
+            self.__setup_camera(cam, nn, nn_node)
 
     def remove_nn(self, sensor_name: str):
         if sensor_name not in self.neural_networks:
@@ -87,17 +107,16 @@ class NNPipeline(StreamingPipeline):
             PipelineQueueType.STILL.get_queue_name(sensor_name)
         )
 
-        cam.preview.unlink(nn_node.input)
         self.pipeline.remove(nn_node)
         del self.neural_networks[sensor_name]
+        self.del_queue(sensor_name, PipelineQueueType.NN)
 
-        if still_queue is not None:
+        if not isinstance(cam, dai.node.MonoCamera) and still_queue is not None:
+            cam.preview.unlink(nn_node.input)
             cam.still.unlink(still_queue.input)
             cam.preview.link(still_queue.input)
             cam.setPreviewSize(cam.getStillSize())
             cam.setNumFramesPool(2, 2, 3, 3, 0)
-        else:
-            print(self.outputs.items())
 
     def remove_sensor(self, sensor):
         self.remove_nn(sensor)
